v 1.0.3
edit/delete categories only for ADMIN role
edit/delete articles only author or ADMIN

+++ src/Controller/BlogController.php
@@ -10,6 +10,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
 
 /**
  * @Route("/blog")
@@ -85,8 +86,12 @@ class BlogController extends AbstractController
     /**
      * @Route("/{id}/edit", name="blog_edit", methods="GET|POST")
      */
-    public function edit(Request $request, Blog $blog): Response
+    public function edit(Request $request, Blog $blog, AuthorizationCheckerInterface $authCheck): Response
     {
+        if (!$authCheck->isGranted($blog->getUser()->getId())) {
+            return $this->redirectToRoute('blog_index');
+        };
+
         $form = $this->createForm(BlogType::class, $blog);
         $form->handleRequest($request);
 
@@ -105,8 +110,12 @@ class BlogController extends AbstractController
     /**
      * @Route("/{id}", name="blog_delete", methods="DELETE")
      */
-    public function delete(Request $request, Blog $blog): Response
+    public function delete(Request $request, Blog $blog, AuthorizationCheckerInterface $authCheck): Response
     {
+        if (!$authCheck->isGranted($blog->getUser()->getId())) {
+            return $this->redirectToRoute('blog_index');
+        };
+
         if ($this->isCsrfTokenValid('delete'.$blog->getId(), $request->request->get('_token'))) {
             $em = $this->getDoctrine()->getManager();
             $em->remove($blog);
